Fix for docker container labels
Given a container image with labels included. When the container is
reloaded the docker module complains about parsing the present
labels. This is because the labels datastructure changed from docker api
version 1.17 to 1.18. It was a list but it became a dictionary.

This commit allows both list and dictionary versions of the api to be
parsed correctly.